drm/hisilicon: dsi: Reset dsi when detect dsi transfer error
This is a workaround for bootup blanking issue.
HDMI can detect dsi transfer error, when this error happened
reseting dsi can bring display to normal.
